Extract PropertyBag from Dynamic and put into its own project
