+++ DotNET.Fundamentals.sln
@@ -117,6 +117,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Concepts.Serialization.Prot
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Immutability", "Source\Immutability\Immutability.csproj", "{00466868-26CA-402D-B99E-C65DA5795E08}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PropertyBags", "Source\PropertyBags\PropertyBags.csproj", "{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PropertyBags", "Specifications\PropertyBags\PropertyBags.csproj", "{28B59332-B699-4FC0-9D12-51829F3EB25C}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -802,6 +806,30 @@ Global
 		{00466868-26CA-402D-B99E-C65DA5795E08}.Release|x64.Build.0 = Release|Any CPU
 		{00466868-26CA-402D-B99E-C65DA5795E08}.Release|x86.ActiveCfg = Release|Any CPU
 		{00466868-26CA-402D-B99E-C65DA5795E08}.Release|x86.Build.0 = Release|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Debug|x64.Build.0 = Debug|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Debug|x86.Build.0 = Debug|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Release|x64.ActiveCfg = Release|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Release|x64.Build.0 = Release|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Release|x86.ActiveCfg = Release|Any CPU
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670}.Release|x86.Build.0 = Release|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Debug|x64.Build.0 = Debug|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Debug|x86.Build.0 = Debug|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Release|x64.ActiveCfg = Release|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Release|x64.Build.0 = Release|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Release|x86.ActiveCfg = Release|Any CPU
+		{28B59332-B699-4FC0-9D12-51829F3EB25C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{AAA185BC-0EA0-4804-9D7B-0E9B1D411070} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
@@ -859,5 +887,7 @@ Global
 		{16A947BA-9B10-4360-B949-841E85009DF2} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
 		{D756E760-E702-46AA-9120-79E3E505CFA7} = {9F86DDBC-3B12-43A9-95F6-C6AD7B6553E2}
 		{00466868-26CA-402D-B99E-C65DA5795E08} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
+		{2BCC600B-7FF7-4FE3-8744-BF0179C3E670} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
+		{28B59332-B699-4FC0-9D12-51829F3EB25C} = {9F86DDBC-3B12-43A9-95F6-C6AD7B6553E2}
 	EndGlobalSection
 EndGlobal
