+++ Source/PropertyBags/PropertyBags.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>Dolittle.PropertyBags</AssemblyName>
+  </PropertyGroup>
+   <ItemGroup>
+    <PackageReference Include="Dolittle.Dynamic" Version="2.0.0-*" />
+    <PackageReference Include="Dolittle.Concepts" Version="2.0.0-*" />
+  </ItemGroup>
+</Project>
