+++ Specifications/PropertyBags/for_ObjectExtensions/for_ToPropertyBag/when_converting_a_dto_with_a_nested_complex_type.cs
@@ -1,8 +1,8 @@
 using System;
-using Dolittle.Dynamic;
 using Machine.Specifications;
+using Dolittle.PropertyBags;
 
-namespace Dolittle.Dynamic.for_ObjectExtensions.for_ToPropertyBag
+namespace Dolittle.PropertyBags.Specs.for_ObjectExtensions.for_ToPropertyBag
 {
     [Subject("ToPropertyBag")]
     public class when_converting_a_dto_with_a_nested_complex_type
