+++ Specifications/PropertyBags/for_ObjectExtensions/for_ToPropertyBag/when_converting_a_simple_dto_with_primitives_as_properties.cs
@@ -1,9 +1,9 @@
 using System;
 using Dolittle.Concepts;
-using Dolittle.Dynamic;
+using Dolittle.PropertyBags;
 using Machine.Specifications;
 
-namespace Dolittle.Dynamic.for_ObjectExtensions.for_ToPropertyBag
+namespace Dolittle.PropertyBags.Specs.for_ObjectExtensions.for_ToPropertyBag
 {
 
     [Subject("ToPropertyBag")]
